refactor(insurance): move the planned-change view mappers into shared/insurance
toPlannedChange, describeChange and toSkipResult are pure view mapping with no
oclif dependency, so they belong with the rest of the insurance code rather
than in the command shell.

They get their own module instead of joining planned-change.ts: rendering a
blob change needs record-update-plan.ts, which already imports
planned-change.ts, so putting them there would close an import cycle.

identityError and applyFailuresError stay behind deliberately -- both build
errors from the command's own cml.import.record-updates message bundle, and
moving them would make a shared module load a command-named bundle.

Pure move: no command name, flag, snapshot or behaviour change, and the suite
stays at 377.
